// src/app/modules/logistica/entrega/monitores/ravex/models/viagem.ts
export interface ILogisticaRavexViagem{
  CD_DIVI_EMPR: number;
  CD_EMPR: number;
  CD_PLAC: string;
  DS_OBSE: string;
  DS_PROD: string;
  DT_EMIS_COLE: Date;
  DT_ESTI_FINA: Date;
  DT_ESTI_INIC: Date;
  DT_FINA_CARR: Date;
  DT_INIC_CARR: Date;
  DT_INIC_PFRI: Date;
  DT_INTE: Date;
  DT_PREV_SAID: Date;
  ID: string;
  ID_COOP: number;
  ID_DEST: number;
  ID_EMBA: number;
  ID_MOTO: number;
  ID_ORIG: number;
  ID_RAVX: number;
  ID_TABE_FRET: number;
  ID_TRAP: number;
  ID_UNID: number;
  ID_VEIC: number;
  ID_VIAG: number;
  IN_COMP_INDI: string;
  IN_INTE: string;
  IN_ORDE_ESPE: boolean;
  KM_ESTI: number;
  NM_MOTO: string;
  NR_COLE: number;
  NR_ORDE_CARR: string
  PESO_BRUT: number;
  PESO_LIQU: number;
  TEMP_MAXI: number;
  TEMP_MINI: number;
  TP_ENTR: string;
  TT_CAIX: number;
  VL_ENTR: number;
}